Remove projects, add README to Lopsided Recovery
Add README files for the Lopsided Recovery project.

Remove Gender Wealth Gap and RCEP Export projects. The former
contains no code documentation and the data are available on
Statistics Indonesia's (BPS) website. The latter contains no
documentation on how to obtain the data. The data files are also
poorly organized.

The main README now tells readers explicitly that this repo is
not an official product of The Jakarta Post.

Some minor updates.