Add description, story table to README
Add a description that tells the reason why this repo exists in
the first place. That is, to host the graphics in my data-driven
stories. Also, add a disclaimer that this is not an official repo.

Add a table of the stories that use the files in this repo. Include
the links to both the story and the repo.